Minor TimIt modification to handle user level timers, optimize timers snapshot, prepare for TimIt IC2D plugin ver 1.01
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@5981 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Core/org/objectweb/proactive/core/util/profiling/TimerProvidable.java
@@ -43,19 +43,19 @@ import org.objectweb.proactive.core.UniqueID;
  * @author vbodnart
  */
 public interface TimerProvidable {
-    public void startTimer(byte timerId, String infos);
+    public void startTimer(int timerId, String infos);
 
-    public void stopTimer(byte timerId, String infos);
+    public void stopTimer(int timerId, String infos);
 
-    public void setTimerValue(byte timerId, long value);
+    public void setTimerValue(int timerId, long value);
 
     public UniqueID getTimerProvidableID();
 
     public void sendResults(String className, String shortUniqueID);
 
-    public void startXAndSkipSendRequest(byte timerId);
+    public void startXAndSkipSendRequest(int timerId);
 
-    public void stopXAndUnskipSendRequest(byte timerId);
+    public void stopXAndUnskipSendRequest(int timerId);
 
     public void stopAll();
